+++ libnavui-androidauto/src/test/java/com/mapbox/androidauto/navigation/CarLocationsOverviewCameraTest.kt
@@ -1,8 +1,8 @@
 package com.mapbox.androidauto.navigation
 
 import com.mapbox.androidauto.car.navigation.CarLocationsOverviewCamera
+import com.mapbox.androidauto.testing.CarAppTestRule
 import com.mapbox.androidauto.testing.MapboxRobolectricTestRunner
-import com.mapbox.common.Logger
 import com.mapbox.maps.CameraOptions
 import com.mapbox.maps.MapSurface
 import com.mapbox.maps.MapboxExperimental
@@ -11,34 +11,20 @@ import com.mapbox.maps.extension.androidauto.MapboxCarMapSurface
 import com.mapbox.maps.plugin.animation.CameraAnimationsPlugin
 import com.mapbox.maps.plugin.animation.camera
 import com.mapbox.navigation.core.MapboxNavigation
-import io.mockk.Runs
 import io.mockk.every
-import io.mockk.just
 import io.mockk.mockk
-import io.mockk.mockkStatic
-import io.mockk.unmockkStatic
 import io.mockk.verify
-import org.junit.After
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertNull
-import org.junit.Before
+import org.junit.Rule
 import org.junit.Test
 
 @OptIn(MapboxExperimental::class)
 class CarLocationsOverviewCameraTest : MapboxRobolectricTestRunner() {
 
-    @Before
-    fun setup() {
-        mockkStatic(Logger::class)
-        every { Logger.e(any(), any()) } just Runs
-        every { Logger.i(any(), any()) } just Runs
-    }
-
-    @After
-    fun teardown() {
-        unmockkStatic(Logger::class)
-    }
+    @get:Rule
+    val carAppTestRule = CarAppTestRule()
 
     @Test
     fun loaded() {
@@ -52,8 +38,9 @@ class CarLocationsOverviewCameraTest : MapboxRobolectricTestRunner() {
         val mapboxCarMapSurface = mockk<MapboxCarMapSurface> {
             every { mapSurface } returns aMapSurface
         }
-        val camera = CarLocationsOverviewCamera(mapboxNavigation)
+        val camera = CarLocationsOverviewCamera()
 
+        carAppTestRule.onAttached(mapboxNavigation)
         camera.onAttached(mapboxCarMapSurface)
 
         verify { mapboxMap.setCamera(any<CameraOptions>()) }
@@ -69,8 +56,9 @@ class CarLocationsOverviewCameraTest : MapboxRobolectricTestRunner() {
         val mapboxCarMapSurface = mockk<MapboxCarMapSurface> {
             every { mapSurface } returns aMapSurface
         }
-        val camera = CarLocationsOverviewCamera(mapboxNavigation)
+        val camera = CarLocationsOverviewCamera()
 
+        carAppTestRule.onAttached(mapboxNavigation)
         camera.onDetached(mapboxCarMapSurface)
 
         assertNull(camera.mapboxCarMapSurface)